rm equity tab, moves equity tab content to wihe page, updates routes
